Rework ARGs vs. ENVs.
Environment variables in images that are imported with FROM are actually not
visible to later commands in the build process. This means we can't
propagate settings from one build stage to the next this way. Instead, we'll
have to resort to repeat the whole ARGs block in every Dockerfile and also
override those ARGs for all 'docker build' invocations. The latter is now
done in the build hook file, which should also be used when building outside
of dockerhub.